+++ helpers/tasks.js
@@ -11,17 +11,17 @@ const Mongo = require(__dirname+'/../db/db.js')
 	, render = require(__dirname+'/render.js')
 	, buildQueue = require(__dirname+'/../queue.js')
 	, gulp = require('gulp')
-	, { rebuild } = require(__dirname+'/../gulpfile.js')
+	, { buildTasks } = require(__dirname+'/../gulpfile.js')
 	, timeDiffString = require(__dirname+'/timediffstring.js');
 
 module.exports = {
 
-	gulp: async () => {
+	gulp: async (options) => {
 		/* TODO: calculate differences in oldsettings vsnewsettings in globalmanagesettings model
 			and send task options with list of tasks instead of always doing all */
-		const label = `gulp tasks [${rebuild.map(x => x.name).join(', ')}] after global config change`;
+		const label = `running gulp tasks [${options.tasks.join(', ')}] after global config change`;
 		const start = process.hrtime();
-		gulp.series(rebuild, () => {
+		gulp.series(options.tasks.map(t => buildTasks[t]), () => {
 			const end = process.hrtime(start);
 			debugLogs && console.log(timeDiffString(label, end));
 		})();